Adding a proper disjoint-set implementation for the spanning-tree function, so we don't have to fake it with a vector of sets anymore.  The result is a 30% speedup.
The dsets are not presently thread-safe, as the "links" member is mutable.  If this ever becomes a problem...  maybe the mutation in STM or something?  We could get rid of the mutability, but it would be at the expense of path compression from find operations.  Will have to experiment on this one.